Add navbar html to separate file
The purpose of having the navbar html in a separate file is to allow for the same navbar code to be applied to all pages.

Each page will run the same JavaScript code which will launch an AJAX GET request to Github and retrieve the html text which will then be displayed on screen.